pyproject.toml
[project]
name = "mae-animal-reconstruct-and-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.3",
]
